requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==4.2
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-newsletter==1.0
ldif3==3.1.1
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
python-card-me==0.9.3
python-dateutil==2.9.0.post0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tzdata==2025.2
unicodecsv==0.14.1
